refactor: move `tesseract::TessBaseAPI` to be thread-local in workers
this commit moves instantiation and usage to the worker.
only creating new tessbase api's is done by the handle class.
previously, for an unknown reason sharing the api would malform
addresses and it would randomly try to write into a protected area
of the memory, resulting in a segmentation fault.
Thanks to Stunt3000 for the help